LayoutTests/fast/dom/Geolocation/coordinates-interface-attributes-expected.txt

Test the attribute handling of the Coordinates interface

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS globalCoordinates.latitude is testSet[currentTestIndex][1][0]
PASS globalCoordinates.longitude is testSet[currentTestIndex][1][1]
PASS globalCoordinates.accuracy is testSet[currentTestIndex][1][2]
PASS globalCoordinates.altitude is testSet[currentTestIndex][1][3]
PASS globalCoordinates.altitudeAccuracy is testSet[currentTestIndex][1][4]
PASS globalCoordinates.heading is testSet[currentTestIndex][1][5]
PASS globalCoordinates.speed is testSet[currentTestIndex][1][6]

PASS globalCoordinates.latitude is testSet[currentTestIndex][1][0]
PASS globalCoordinates.longitude is testSet[currentTestIndex][1][1]
PASS globalCoordinates.accuracy is testSet[currentTestIndex][1][2]
PASS globalCoordinates.altitude is testSet[currentTestIndex][1][3]
PASS globalCoordinates.altitudeAccuracy is testSet[currentTestIndex][1][4]
PASS globalCoordinates.heading is testSet[currentTestIndex][1][5]
PASS globalCoordinates.speed is testSet[currentTestIndex][1][6]

PASS globalCoordinates.latitude is testSet[currentTestIndex][1][0]
PASS globalCoordinates.longitude is testSet[currentTestIndex][1][1]
PASS globalCoordinates.accuracy is testSet[currentTestIndex][1][2]
PASS globalCoordinates.altitude is testSet[currentTestIndex][1][3]
PASS globalCoordinates.altitudeAccuracy is testSet[currentTestIndex][1][4]
PASS globalCoordinates.heading is testSet[currentTestIndex][1][5]
PASS globalCoordinates.speed is testSet[currentTestIndex][1][6]

PASS globalCoordinates.latitude is testSet[currentTestIndex][1][0]
PASS globalCoordinates.longitude is testSet[currentTestIndex][1][1]
PASS globalCoordinates.accuracy is testSet[currentTestIndex][1][2]
PASS globalCoordinates.altitude is testSet[currentTestIndex][1][3]
PASS globalCoordinates.altitudeAccuracy is testSet[currentTestIndex][1][4]
PASS globalCoordinates.heading is testSet[currentTestIndex][1][5]
PASS globalCoordinates.speed is testSet[currentTestIndex][1][6]

PASS globalCoordinates.latitude is testSet[currentTestIndex][1][0]
PASS globalCoordinates.longitude is testSet[currentTestIndex][1][1]
PASS globalCoordinates.accuracy is testSet[currentTestIndex][1][2]
PASS globalCoordinates.altitude is testSet[currentTestIndex][1][3]
PASS globalCoordinates.altitudeAccuracy is testSet[currentTestIndex][1][4]
PASS globalCoordinates.heading is testSet[currentTestIndex][1][5]
PASS globalCoordinates.speed is testSet[currentTestIndex][1][6]

PASS successfullyParsed is true

TEST COMPLETE
